requirements.txt
Flask
werkzeug
pdfplumber
pydantic~=2.0
langchain-google-genai==2.1.3
python-dotenv
requests
streamlit
pandas
flask-cors
google-ai-generativelanguage==0.6.17